requirements.txt
beautifulsoup4==4.9.0
Flask==1.1.2
geopy==1.21.0
python-telegram-bot==12.7
requests==2.23.0
SQLAlchemy==1.3.16
xmltodict==0.12.0
psycopg2==2.8.5
gunicorn==20.0.4
